+++ src/chart/index.js
@@ -26,5 +26,5 @@ export function createChart(container, options) {
   return new Chart(container, options)
 }
 
-export const version = '0.6.2'
+export const version = '0.7.0'